【カードマスタ】【Lambda】【データINSERT用】Neuron関連 #31
- 設計書の修正
- データの型を修正
- コメントの見直し

+++ code/cardApiCall/infoInsert/usecase/neuron.go
@@ -10,24 +10,29 @@ import (
 	"github.com/samber/do"
 )
 
+// neuronUseCaseImpl は、NeuronUseCaseの実装です。
 type neuronUseCaseImpl struct {
 	*useCase
 }
 
+// NeuronUseCase は、NeuronUseCaseのインターフェースです。
 type NeuronUseCase interface {
 	UseCase
-	GetCardInfo(ctx context.Context, cardID int) (NeuronExtractedData, error)
+	GetCardInfo(ctx context.Context, cardID int64) (NeuronExtractedData, error)
 }
 
+// NewNeuronUseCase は、NeuronUseCaseのコンストラクタです。
 func NewNeuronUseCase(i *do.Injector) (NeuronUseCase, error) {
 	return NewUseCase(i, func(u *useCase) NeuronUseCase {
 		return &neuronUseCaseImpl{u}
 	})
 }
 
+// emptyFunc は、空の関数です。
 func (n *neuronUseCaseImpl) emptyFunc() {
 }
 
+// NeuronExtractedData は、NeuronUseCaseの抽出データです。
 type NeuronExtractedData struct {
 	CardNameEn     string
 	CardNameJa     string
@@ -35,7 +40,8 @@ type NeuronExtractedData struct {
 	PendulumTextJa string
 }
 
-func (n *neuronUseCaseImpl) GetCardInfo(ctx context.Context, cardID int) (NeuronExtractedData, error) {
+// GetCardInfo により、NeuronUseCaseを使ってカードの情報を取得する
+func (n *neuronUseCaseImpl) GetCardInfo(ctx context.Context, cardID int64) (NeuronExtractedData, error) {
 	htmlGetter := htmlget.NewNeuronHtmlGetter()
 	results, err := htmlGetter.VisitSite(ctx, fmt.Sprintf(htmlget.BASE_URL_FORMAT, cardID))
